GDNative: Make godot_int an int64_t
Redo of the change in cf8c679a23b21d6c6f29cba6a54eaa2eed88bf92
that caused a build issue, with the extra change needed to
'godot_arvr_blit'.

+++ modules/gdnative/include/gdnative/gdnative.h
@@ -127,7 +127,7 @@ typedef bool godot_bool;
 
 /////// int
 
-typedef int godot_int;
+typedef int64_t godot_int;
 
 /////// real
